+++ src/cpac/utils/__init__.py
@@ -1,8 +1,9 @@
 from .checks import check_version_at_least
-from .utils import LocalsToBind, PermissionMode, Volume, Volumes
+from .utils import LocalsToBind, PermissionMode, Volume, Volumes, get_project_root
 
 __all__ = [
     "check_version_at_least",
+    "get_project_root",
     "LocalsToBind",
     "PermissionMode",
     "Volume",